// src/models/npcs/species/beasts/traits/combat/tags.ts
export type beast__combat_trait_tags =
  | 'chameleon'
  | 'armored'
  | 'packs'
  | 'keen_senses'
  | 'vicious'
  | 'stealth'
  | 'venomous'
  | 'constrictor'
  | 'odor'
  | 'unpalatable'
  | 'quills'
  | 'agile'
  | 'breath'
  | 'electric'
  | 'petrify'
  | 'regeneration'
  | 'invisibility'
  | 'teleportation'
  | 'illusions'